chore: change packages structure
