+++ app/src/main/kotlin/model/WeightedUndirectedGraph.kt
@@ -1,18 +1,19 @@
-package model.graph
+package model
 
-class WeightedDirectedEdge<D>(vertex1: Vertex<D>, vertex2: Vertex<D>, val weight: Int) :
-    DirectedEdge<D>(vertex1, vertex2)
+import model.abstractGraph.Vertex
+import model.edges.WeightedUndirectedEdge
 
-class WeightedDirectedGraph<D> : DirectedGraph<D>() {
-    fun addEdge(vertex1: Vertex<D>, vertex2: Vertex<D>, weight: Int): WeightedDirectedEdge<D>? {
+class WeightedUndirectedGraph<D> : UndirectedGraph<D>() {
+    fun addEdge(vertex1: Vertex<D>, vertex2: Vertex<D>, weight: Int): WeightedUndirectedEdge<D>? {
         if (vertex1 == vertex2 ||
             vertex1 !in adjacencyMap.keys ||
             vertex2 !in adjacencyMap.keys
         ) return null
 
-        val newEdge = WeightedDirectedEdge(vertex1, vertex2, weight)
+        val newEdge = WeightedUndirectedEdge(vertex1, vertex2, weight)
         edges.add(newEdge)
         adjacencyMap[vertex1]?.add(vertex2)
+        adjacencyMap[vertex2]?.add(vertex1)
 
         return newEdge
     }
