+++ app/src/main/kotlin/model/abstractGraph/Vertex.kt
@@ -1,3 +1,3 @@
-package model.graph
+package model.abstractGraph
 
 class Vertex<D>(val id: ULong, val data: D)
